client: refactor media related stuff into another package
This makes for more logical code, e.g. media.Orientation*, instead
of client.Orientation* and etc.
