+++ client/media/post.go
@@ -1,4 +1,4 @@
-package client
+package media
 
 import (
 	"errors"
